Django_Backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
mysqlclient==2.2.7
numpy==2.4.0
scikit-learn==1.8.0
pandas==2.3.3
django-jet-reboot

